Remove rootfs volume from cAdvisor due to bug
When using the rootfs volume recommended in the README getting
started guide, cAdvisor prevents containers that it is monitoring
from being cleaned up properly upon stopping.

This patch removes that volume, which is supposed to allow cAdvisor
to degrade the stats that it tracks gracefully and also fixing the
problem until a better solution is found.

Relevant issue link: https://github.com/google/cadvisor/issues/771