Move to Vagrant for vm management
Previously these scripts required you to manually spin up your own
vm with virtualbox and configure it with an ssh key and whatnot. Now
the Vagrantfile has all of this version controlled, so you can just
do `vagrant up` to get the vm bootstrapped and ready to deploy to.

Also:

- renamed 'scripts' directory to 'ansible'
- upgraded cAdvisor to 0.24.1
- upgraded docker-gc to latest master branch